BFF-625 Moved sign up page to registration folder
