+++ lib/features/registration/widgets/widgets.dart
@@ -1,0 +1,1 @@
+export 'registration_app_bar.dart';
